Varios cambios

+++ view/modules/admin/readAsignacion.php
@@ -81,7 +81,9 @@
           <td><?php echo $row['user_name'] ?></td>
           <td><?php echo $row['asig_fecha']; ?></td>
           <td><a href="?detalle=<?php echo $row['asig_id']; ?>&c=admin&a=DetalleAsignacion" class="btn waves-effect waves-light blue-grey darken-2 tooltipped" data-position="left" data-tooltip="Modificar asignación"><i class="small material-icons">update</i></a></td>
-          <td><a href="?id=<?php echo $row['asig_id']; ?>&c=admin&a=ReadDeleteAsignacion" class="btn waves-effect waves-light red darken-1 tooltipped" data-position="top" data-tooltip="Eliminar asignacion"><i class="small material-icons">delete</i></a></td>
+          <?php if($_SESSION["user"]["user"] != "2"){ ?>
+            <td><a href="?id=<?php echo $row['asig_id']; ?>&c=admin&a=ReadDeleteAsignacion" class="btn waves-effect waves-light red darken-1 tooltipped" data-position="top" data-tooltip="Eliminar asignacion"><i class="small material-icons">delete</i></a></td>
+          <?php } ?>
         </tr>
       </tbody>
     </table>
@@ -123,7 +125,9 @@
           <td><?php echo $data->user_name; ?></td>
           <td><?php echo $data->asig_fecha; ?></td>
           <td><a href="?detalle=<?php echo $data->asig_id; ?>&c=admin&a=DetalleAsignacion" class="btn waves-effect waves-light blue-grey darken-2 tooltipped" data-position="left" data-tooltip="Modificar asignación"><i class="small material-icons">update</i></a></td>
-          <td><a href="?id=<?php echo $data->asig_id; ?>&c=admin&a=ReadDeleteAsignacion" class="btn waves-effect waves-light red darken-1 tooltipped" data-position="top" data-tooltip="Eliminar asignación"><i class="small material-icons">delete</i></a></td>
+          <?php if($_SESSION["user"]["rol"] != "2"){ ?>
+            <td><a href="?id=<?php echo $data->asig_id; ?>&c=admin&a=ReadDeleteAsignacion" class="btn waves-effect waves-light red darken-1 tooltipped" data-position="top" data-tooltip="Eliminar asignación"><i class="small material-icons">delete</i></a></td>
+          <?php } ?>
         </tr>
       </tbody>
